Move PluginHandler to separate file
