+++ staging/src/k8s.io/kubectl/pkg/cmd/cmd.go
@@ -20,11 +20,7 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
 	"strings"
-	"syscall"
 
 	"github.com/spf13/cobra"
 
@@ -165,141 +161,6 @@ func NewDefaultKubectlCommandWithArgs(o KubectlOptions) *cobra.Command {
 	}
 
 	return cmd
-}
-
-// IsSubcommandPluginAllowed returns the given command is allowed
-// to use plugin as subcommand if the subcommand does not exist as builtin.
-func IsSubcommandPluginAllowed(foundCmd string) bool {
-	allowedCmds := map[string]struct{}{"create": {}}
-	_, ok := allowedCmds[foundCmd]
-	return ok
-}
-
-// PluginHandler is capable of parsing command line arguments
-// and performing executable filename lookups to search
-// for valid plugin files, and execute found plugins.
-type PluginHandler interface {
-	// exists at the given filename, or a boolean false.
-	// Lookup will iterate over a list of given prefixes
-	// in order to recognize valid plugin filenames.
-	// The first filepath to match a prefix is returned.
-	Lookup(filename string) (string, bool)
-	// Execute receives an executable's filepath, a slice
-	// of arguments, and a slice of environment variables
-	// to relay to the executable.
-	Execute(executablePath string, cmdArgs, environment []string) error
-}
-
-// DefaultPluginHandler implements PluginHandler
-type DefaultPluginHandler struct {
-	ValidPrefixes []string
-}
-
-// NewDefaultPluginHandler instantiates the DefaultPluginHandler with a list of
-// given filename prefixes used to identify valid plugin filenames.
-func NewDefaultPluginHandler(validPrefixes []string) *DefaultPluginHandler {
-	return &DefaultPluginHandler{
-		ValidPrefixes: validPrefixes,
-	}
-}
-
-// Lookup implements PluginHandler
-func (h *DefaultPluginHandler) Lookup(filename string) (string, bool) {
-	for _, prefix := range h.ValidPrefixes {
-		path, err := exec.LookPath(fmt.Sprintf("%s-%s", prefix, filename))
-		if shouldSkipOnLookPathErr(err) || len(path) == 0 {
-			continue
-		}
-		return path, true
-	}
-	return "", false
-}
-
-func Command(name string, arg ...string) *exec.Cmd {
-	cmd := &exec.Cmd{
-		Path: name,
-		Args: append([]string{name}, arg...),
-	}
-	if filepath.Base(name) == name {
-		lp, err := exec.LookPath(name)
-		if lp != "" && !shouldSkipOnLookPathErr(err) {
-			// Update cmd.Path even if err is non-nil.
-			// If err is ErrDot (especially on Windows), lp may include a resolved
-			// extension (like .exe or .bat) that should be preserved.
-			cmd.Path = lp
-		}
-	}
-	return cmd
-}
-
-// Execute implements PluginHandler
-func (h *DefaultPluginHandler) Execute(executablePath string, cmdArgs, environment []string) error {
-	// Windows does not support exec syscall.
-	if runtime.GOOS == "windows" {
-		cmd := Command(executablePath, cmdArgs...)
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		cmd.Stdin = os.Stdin
-		cmd.Env = environment
-		err := cmd.Run()
-		if err == nil {
-			os.Exit(0)
-		}
-		return err
-	}
-
-	// invoke cmd binary relaying the environment and args given
-	// append executablePath to cmdArgs, as execve will make first argument the "binary name".
-	return syscall.Exec(executablePath, append([]string{executablePath}, cmdArgs...), environment)
-}
-
-// HandlePluginCommand receives a pluginHandler and command-line arguments and attempts to find
-// a plugin executable on the PATH that satisfies the given arguments.
-func HandlePluginCommand(pluginHandler PluginHandler, cmdArgs []string, minArgs int) error {
-	var remainingArgs []string // all "non-flag" arguments
-	for _, arg := range cmdArgs {
-		if strings.HasPrefix(arg, "-") {
-			break
-		}
-		remainingArgs = append(remainingArgs, strings.Replace(arg, "-", "_", -1))
-	}
-
-	if len(remainingArgs) == 0 {
-		// the length of cmdArgs is at least 1
-		return fmt.Errorf("flags cannot be placed before plugin name: %s", cmdArgs[0])
-	}
-
-	foundBinaryPath := ""
-
-	// attempt to find binary, starting at longest possible name with given cmdArgs
-	for len(remainingArgs) > 0 {
-		path, found := pluginHandler.Lookup(strings.Join(remainingArgs, "-"))
-		if !found {
-			remainingArgs = remainingArgs[:len(remainingArgs)-1]
-			if len(remainingArgs) < minArgs {
-				// we shouldn't continue searching with shorter names.
-				// this is especially for not searching kubectl-create plugin
-				// when kubectl-create-foo plugin is not found.
-				break
-			}
-
-			continue
-		}
-
-		foundBinaryPath = path
-		break
-	}
-
-	if len(foundBinaryPath) == 0 {
-		return nil
-	}
-
-	// invoke cmd binary relaying the current environment and args given
-	if err := pluginHandler.Execute(foundBinaryPath, cmdArgs[len(remainingArgs):], os.Environ()); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 // NewKubectlCommand creates the `kubectl` command and its nested children.
